+++ docs/Configuration.md
@@ -156,7 +156,6 @@ nav_order: 15
 | spark.gluten.sql.orc.charType.scan.fallback.enabled                | true              | Force fallback for orc char type scan.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     |
 | spark.gluten.sql.parquet.maxmin.index                              | false             | Enable row group max min index for parquet file scan                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
 | spark.gluten.sql.removeNativeWriteFilesSortAndProject              | true              | When true, Gluten will remove the vanilla Spark V1Writes added sort and project for velox backend.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         |
-| spark.gluten.sql.rewrite.castArrayToString                         | true              | When true, rewrite `cast(array as String)` to `concat('[', array_join(array, ', ', null), ']')` to make velox support offload.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | spark.gluten.sql.rewrite.dateTimestampComparison                   | true              | Rewrite the comparision between date and timestamp to timestamp comparison.For example `from_unixtime(ts) > date` will be rewritten to `ts > to_unixtime(date)`                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            |
 | spark.gluten.sql.scan.fileSchemeValidation.enabled                 | true              | When true, enable file path scheme validation for scan. Validation will fail if file scheme is not supported by registered file systems, which will cause scan  operator fall back.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        |
 | spark.gluten.sql.substrait.plan.logLevel                           | DEBUG             |
